+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 **unreleased**
 
+**v0.0.1**
 
 - Added row and column wrappers for indxl2g
 - Added a wrapper for pdgemr2d